feat(admin): split users dashboard into onboarding and retention hubs
Extract onboarding funnel/sources/cohorts and retention trials/churn/inactive
pages from the monolithic users dashboard; keep users.vue as a redirect.
