+++ messages/en.json
@@ -241,6 +241,34 @@
   "admin-credits-select-org": "Select Organization",
   "admin-credits-title": "Grant Credits",
   "admin-dashboard": "Admin Dashboard",
+  "admin-pulse": "Pulse",
+  "admin-onboarding": "Onboarding",
+  "admin-onboarding-funnel": "Funnel",
+  "admin-onboarding-sources": "Sources",
+  "admin-onboarding-cohorts": "Cohorts",
+  "admin-onboarding-cohorts-description": "Stuck onboarding stages derived from the live funnel. Use this surface to prioritize who needs attention.",
+  "admin-onboarding-cohorts-hint": "This is the action surface for stuck onboarding — counts come from the same onboarding funnel stats as the Funnel tab.",
+  "admin-onboarding-stuck-app-no-channel": "App but no channel",
+  "admin-onboarding-stuck-app-no-channel-description": "Organizations that created an app but never created a channel",
+  "admin-onboarding-stuck-channel-no-bundle": "Channel but no bundle",
+  "admin-onboarding-stuck-channel-no-bundle-description": "Organizations with a channel that never uploaded a bundle",
+  "admin-onboarding-stuck-bundle-no-device": "Bundle but no production device",
+  "admin-onboarding-stuck-bundle-no-device-description": "Organizations with a bundle that have no production plugin device yet",
+  "admin-onboarding-stuck-device-no-download": "Device but no update download",
+  "admin-onboarding-stuck-device-no-download-description": "Organizations with a production device that have not completed an update download",
+  "admin-product": "Product",
+  "admin-retention": "Retention",
+  "admin-retention-trials": "Trials",
+  "admin-retention-churn": "Churn",
+  "admin-retention-inactive": "Inactive",
+  "admin-retention-inactive-description": "Paid organizations with zero MAU or no uploads in the selected period.",
+  "admin-customers": "Customers",
+  "admin-revenue": "Revenue",
+  "admin-revenue-overview": "Overview",
+  "admin-revenue-upgrades": "Upgrades",
+  "admin-revenue-risk": "Risk",
+  "admin-platform": "Platform",
+  "admin-platform-capacity": "Capacity",
   "admin-organizations": "Organizations",
   "admin-users-email-type-breakdown": "Email Type Breakdown",
   "admin-users-email-type-breakdown-description": "Registration mix in the selected period, split between professional, personal, and disposable email domains.",
