+++ src/pages/admin/dashboard/platform/debug.vue
