+++ src/pages/admin/dashboard/product/cli.vue
