+++ src/pages/admin/dashboard/product/plugins.vue
